Download multiple CSVs from `gs://` in parallel (UNFINISHED)
We can now copy data from CSV files, to PostgreSQL, to Google Cloud
Storage, and back to CSV files.

First, CSV to PostgreSQL:

    dbcrossbar cp \
        --if-exists=overwrite \
        --schema=postgres-sql:table.sql \
        csv:example.csv \
        'postgresql://localhost:5432/test#example'

Then, PostgreSQL to Google Cloud Storage:

    dbcrossbar cp \
        --if-exists=overwrite \
        'postgresql://localhost:5432/test#example' \
        gs://my-bucket/csvs/

    dbcrossbar cp \
        --schema=postgres-sql:table.sql \
        gs://my-bucket/csvs/ \
        csv:out/

This should create `out/example.csv` with the same contents as the
original `example.csv`.

This is doubly async: It runs an async recursive `gsutil ls` on the
underlying cloud bucket, and then it runs multiple async `gsutil cp
gs://... -` up to the `buffer(...)` limit.

This passes the integration tests successfully, but it doesn't yet
handle async `gsutil` failures in a really convincing fashion, because
they currently happen in disconnected background `tokio` workers that
don't have anybody to report errors to.