Deep async magic: Handle cancellation on error
When we're running, we usually have many simultaneous background and
foreground workers. But if, for example, a background `gsutil cp` fails
somewhere, we want to shut _everything_ down in a reasonably timely
manner.

This code attempts to implement that. We make a rule that all
`tokio::spawn` invocations (which start an independent worker process)
must be done using a `Context` type which forwards any errors to a
channel. We create a future that waits for either an error to be sent on
this channel, or for all senders to be dropped.

This allows us to create a `worker_fut` and use the `Future::join`
command to combine multiple futures into a single future that fails as
soon as _any_ input fails.

A failure will tear down everything in the system _except_ for spawned
background workers. But these will generally error out relatively
quickly when they try to send output and get hit with a "closed pipe"
error, either from `tokio` or from the OS, because their consumers will
be gone.