Move parallelism control out of `write_local_data`
Before, `write_local_data` was in charge of managing parallelism
internally, using something like a `BoxStream<BoxFuture<()>>` and the
`.buffered(n)` function to resolve up to `n` of those `BoxFuture<()>`
values in parallel.

But I want to move this logic out to the top level, so that we can
globally control the degree of overall parallelism based on (1) what the
input driver can support, (2) what the output driver can support,
and (3) an explicit override by the user.