Load data from `gs://` to BigQuery
This particular case is a bit interesting, because unlike all our other
implementations, it never actually creates CSV streams in memory, but
instead does all the work on an external system.

To do this, we need to use `Any` and `downcast_ref` to turn a `&dyn
Locator` back into a concrete `&GsLocator`. This is definitely a less
well-developed corner of Rust, and it's surprisingly hard to get
working. Normally we'd avoid using `downcast_ref` on a `trait Locator`
in favor of an `enum Locator` (much more idiomatic), but then `Locator`s
would become a closed type, which is the wrong design, because we
_really_ want to be able to add new backends at will. So using
`downcast_ref` is a compromise between two different designs, either of
which would be more idiomatic.