Spike remaining `FORMAT BINARY` types (except `decimal`)
This covers all the current `dbcrossbar` types except `decimal`, which
could most easily be supported by porting the actual PostgreSQL decimal
parser from C to Rust. (The `decimal` type is all about perfect lossless
representation, and the binary representation of `decimal` is not well
documented. So the easiest way to ensure we produce the correct binary
representation is to parse string inputs in exactly the same way
PostgreSQL does.)