Refactor all BigQuery-related code
This refactoring purges most of the weird code structures left over from
the original `schemaconv` BigQuery support (but not the
PostgreSQL-related oddities, which still remain to be done). We switch
from a "random grab bag of functions" model to "lots of fairly
predictable types with methods" model, though not—strictly
speaking—anything involving inheritance.