postgres: Begin building real PostgreSQL type structures
Up until now, we've never really had a data structure that corresponds
to real PostgreSQL types. Instead, we've just converted to and from the
portable `DataType` immediately.

This patch adds a real `PgDataType` enum which directly represents
native PostgreSQL types, and which can be converted to and from the
cross-database `DataType`.