Postgres: Work around unvalidateable certs
PostgreSQL's `libpq` has an insecure default where it will allow an SSL
connection even when it can't validate the SSL cert. And many real-world
PostgreSQL databases have non-validateable certs. This leaves them wide
open to man-in-the-middle attacks.

We use `diesel` for most of our PostgreSQL support, and `diesel`
internally uses `libpq`, so it inherits this default behavior.

However, for actually moving data, we use the Rust `postgres` library,
which (being a conscientious Rust library), always insists on fully
validating SSL connections.

With this patch, there are now three recommended ways to connect to a
PostgreSQL database:

- `$URL`: Validate SSL if present, otherwise fall back to an insecure
  connection.
- `$URL?sslmode=verify-full`: Always use SSL, and validate it fully.
- `$URL?sslmode=disable`: Never use SSL.

Note that we do not support `sslmode=require` or `sslmode=verifyca`, and
that our default mode always validates certs, unlike `sslmode=prefer`.