v0.0.12: Provide a workaround for PostgreSQL SSL quirks (& BigQuery fix)
PostgreSQL's `libpq` has an insecure default where it will allow an SSL
connection even when it can't validate the SSL cert. And many real-world
PostgreSQL databases have non-validateable certs. This leaves them wide
open to man-in-the-middle attacks.

With this release, there are now three recommended ways to connect to a
PostgreSQL database:

- `$URL`: Validate SSL if present, otherwise fall back to an insecure
  connection.
- `$URL?sslmode=verify-full`: Always use SSL, and validate it fully.
- `$URL?sslmode=disable`: Never use SSL.

Note that we do not support `sslmode=require` or `sslmode=verifyca`, and
that our default mode always validates certs, unlike `sslmode=prefer`.

We also improve our parsing of BigQuery JSON schemas.