Update PostgreSQL for OpenSSL 1.1 & tokio
To get a version of PostgreSQL compatible with OpenSSL 1.1, we need to
use 0.16.0-rc.1. But that version of `postgres` is just a wrapper around
`tokio-postgres`, which is what we'd prefer to be using anyway. So we
update to a new nightly build and convert a bunch of sync code to
async (yay!).

It's entirely possible that we'll need to revist PostgreSQL SSL mode
specifications one more time before the final release, to get `postgres`
and `libpq` back on the same page.

Many thanks to @n1ywb for pairing with me on several big pieces of this.