Remove `Compat` wrapper from `BoxFuture<T>`
Our core future type is `BoxFuture<T>`, which is currently a
`tokio`-compatible future. This patch converts it to a
`std::future::Future` by removing that wrapper. This involves a
converting a lot of `.boxed().compat()` to `.boxed()`, and inserting the
`.compat()` calls in other places.

This means that we're getting ever-closer to a world where we use
`std::future::Future` everywhere. We're just waiting for `tokio` and the
surrounding ecosystem to make the jump, which may take a while. And for
`.await` to stablize.