Concatenate CSV streams (not exposed yet)
This patch adds what should be a high-performance CSV concatenator,
which should allow doing things like "download multiple CSV files as a
single CSV file." In many cases, this isn't an especially good
idea--it's often better to work with sets of CSV files for scalability
reasons--but there are cases where this is necessary.

This patch factors out a bunch of reusable code. It also adds quite a
bit of new test machinery so that we can test async code without having
to test it via the CLI.

(We also fix the the CLI logging routines to use `stderr` instead of
`stdout`. Not sure what that was like that.)