BigQuery: Only allow insert on REQUIRED columns (fixes #43)
It turns out that upsert and `NULL` do not play nicely
together. BigQuery joins support neither `IS NOT DISTINCT FROM` nor
compound tests using `IS NOT NULL`. So the best decision here is to only
allow joins on columns which are known to be `REQUIRED` (which maps to
`NOT NULL` columns using SQL table definitions).