Fix deadlock in `copy_reader_to_stream`
Inside an `async` block, we're support to call `my_future.await`, and
_never_ use the sync-mode blocking equivalent `my_future.wait()`.
Apparently, doing this can deadlock the entire `tokio` runtime, or at
least the portions touching file I/O.

Fixes #16. Fixes #46.

We've also submitted a "lint" proposal to the clippy project which would
automatically detect errors of this class in the future:
https://github.com/rust-lang/rust-clippy/issues/4427