WIP: Upload data to BigML via S3
This is fairly clunky to use on the CLI because it requires a
`--temporary=s3://...` argument and doesn't work with
`AWS_SESSION_TOKEN`. I adapted some pre-existing code that we had for
URL signing and column data type correction.

I also moved around various code for creating temporary storage on S3
and GCS, just to make things a bit more logical.