WIP: BigML: Fetch dataset schemas (+ more async)
This was more work than I expected, because the `Locator::schema` and
`Locator::write_schema` methods were still written using synchronous
Rust (unlike `local_data`, `write_local_data` and `write_remote_data`).
There wasn't any principled reason for this; it was purely historic.

But this meant that it was difficult to use the async `bigml` crate to
fetch schema information for BigML datasets. So I needed to convert the
main `Locator` APIs to be fully async, which required a bunch of
changes.