bigml: Add `--to-arg` keys `name` and `optype_for_text`
These can be used to set the BigML name of the sources and datasets we
create, and to specify the BigML optype to use for "text" columns.

We add a bunch of nice new parsing machinery for `DriverArgs`, to allow
deserializing `--to-arg` into a `serde`-compatible struct type, which
makes everything very nice and Rust-like.