v0.2.0: Add many new backends, fix GCS & S3 stalls
This release of `dbcrossbar` adds many new backends. The current list
is:

- bigml
- bigquery
- bigquery-schema
- csv
- dbcrossbar-schema
- gs
- postgres
- postgres-sql
- redshift
- s3

For an up-to-date list, run `dbcrossbar features`. For information on a
specific driver, run `dbcrossbar features $DRIVER`.

This release also fixes a long-standing bug with `gs://` and `s3://`
inputs that could cause `dbcrossbar`'s internal pipelines to deadlock
and stop transfering data.