Postgres: Append to tables with legacy JSON columns
If `dbcrossbar` creates a new PostgreSQL table, it will always use
`jsonb` columns instead of `json` columns. But previously, if there was
a pre-existing table with legacy `json`, then `dbcrossbar` would fail to
insert data.

The fix is read the _actual_ table metadata out of the database, and use
that to perform our conversion from CSV to PostgreSQL BINARY format.
This allows us to encode data as BINARY `json`. We need to be careful to
make sure our table columns are correctly aligned, in case a
pre-existing destination table had columns in a different order, or
whatever.