Switch from `bitflag` to `enumset`
The `bitflag` crate is used to define sets of bits (0x01, 0x02, 0x04,
etc.) which can be ORed together and used to represent small sets. It's
very efficient, because a set can be represented as a single integer.

The `enumset` crate does pretty much the same thing, except using a
regular Rust `enum`. It's cleaner and offers fewer opportunities to mess
up.