v0.2.4: Fix `bigml` to `bigquery` transfers
The `bigml` driver was including a leading "dataset/" in all stream
names, which the `bigquery` driver couldn't handle. This is a quick fix
that just removes the leading "dataset/", but we'll need to do further
work on the `bigquery` and `gs` drivers to handle names with "/" (which
represent directory structure).