postgres: Add `character` data type
This is simple case: We just need to add a new grammar rule to the SQL
parser, and to the code that parses `information_schema.columns`. From
there, we can safely translate it to `text` immediately, and the tests
all pass. No need to change the Postgres BINARY encoder, because
apparently `character` is encoded the same way as `varchar` and `text`
on the wire.

This has not been tested with Redshift, but if a given PostgreSQL type
works at all with Redshift, then we usually don't need to do more than
this.