bigquery: Quote each part of table name (fixes #79)
This is weird, because we've been using the old quoting code for a quite
a while now, and it has been working just fine in lots of different
circumstances. But currently it's failing. Happily this should be an
easy fix. We just need to clean up some crufty old `schemaconv` code and
make `DottedTableName` a bit smarter.