Require exact output for select scalar types
We now require all drivers to output the following types as exact byte
sequences:

- null: The empty string.
- bool: "t" or "f".
- date: "YYYY-MM-DD"
- int types: As expected.
- text types: Exact input text.
- Timestamps w/ time zones: "yyyy-mm-ddThh:mm:ssZ".
- Timestamps w/o time zones: "yyyy-mm-ddThh:mm:ss".
- UUIDs: Lowercase hex letters & dashes

Other types have not yet been standardized.

PostgreSQL timestamp handling is a bit gross because we share code
between PostgreSQL and Redshift, and Redshift doesn't allow temporary
functions. So we need to inline all the necessary code, which is
actually fairly complex.