postgres: Upsert legacy `json` columns (fixes #85)
The key trick here is that we need to fetch any existing schema _before_
calling `prepare_table`. This whole system needs to be re-designed as
part of the "shortcuts" system some day. We also need to inspect a
number of other functions in other drivers for similar problems. Again,
there's a missing abstraction here.

To make this work, we encapsulated the pg_catalog reading logic and
forced all callers to go through the same API. We also fixed an issue
where catalog lookups were being done synchronously inside async code.