Port to `tokio` 0.2
This is a big, complicated patch which updates us from `tokio` 0.1 to
`tokio` 0.2. This requires us to update many of our networking-related
dependencies, which also make similar jumps in this release. And all
these updates, taken together, allow us to finally switch from Rust
nightly-2019-09-05 (no earlier, no later) to current stable Rust.

For more information about async Rust, see:
https://rust-lang.github.io/async-book/index.html

Some major categories of changes include:

**Only one kind of futures, and no more `compat`**

The `.compat()` method was used to translate between the old
`tokio::Future` and `std::futures::Future`. Now that tokio uses standard
futures, this is no longer needed.

**New stream type**

Stream types are no longer specified as:

    Stream<Item = Bytes, Error = Error>

They are now specified as:

    Stream<Item = Result<Bytes, Error>>

This has a lot of ripple effects throughout the API, replacing `.map`
with `.map_ok`, etc.

**Iterating over streams**

We replace `stream.into_future()` with `stream.next()`, a very nice API
cleanup that appears throughout the code. **Please check these callsites
carefully to make sure they preserve the original logic.**

**Stream boxing**

Streams may now also be boxed (placed onto the heap) using the
`stream.boxed()` method, which also "pins" the stream so that it can't
be copied out of the box. This is necessary in case the stream object
contains an `async` function. For more background, see the async Rust
book. This usually replaces `Box::new`, and it results in many more
types with `Pin<Box<..>>` instead of `Box<..>`.

**Relacing `sender = sender.send(..)` with `sender.send(..)`**

Many tokio APIs used to consume their `sender`, `receiver`, `stream` or
`sink` arguments, and then return a replacement for that value. Most of
these methods now take their arugment by `&mut self` instead of `self`.

This was a source of many test suite bugs, because it means that we no
longer close a `sender` or `AsyncWrite` implementation implicitly, and
in some cases we now need to do it manually to prevent hangs.

**No more `tokio_process::CommandExt`**

Now we use `tokio::process::Command` directly, which changes a few minor
APIs.

**No more `format!` on the same line as `await`**

When we call `format!`, Rust creates some temporary data structures that
will live until the end of the current "line" (i.e., the next
semi-colon). But if that line also contains an `.await`, Rust will
refuse to compile, because the temporary structures are not marked as
`Send`-safe, and cannot be moved between threads during the `.await`.

**`tokio-postgres` API changes**

Our Postgres API has changed in a lot of little ways.

There are also other changes, but these are the big ones.