Replace `wkb` with `postgis` for licensing reasons
The `wkb` library converted GIS data types (lat/lon, etc.) into a
PostgreSQL-compatible binary type. However, it was under the AGPL,
which restricted what could be done with `dbcrossbar`.

This patch replaces `wkb` with the newer `postgis` library, and provides
the necessary glue code to make it work. We also delete the old
pg_copy_binary spike, because it would take extra work to update it.

Finally, we replace `cargo audit` with `cargo deny`, and we configure
it to check for security alerts and to enforce a project-wide license
policy, in order to prevent any licensing regressions. Interestingly,
dbcrossbar is now entirely under permissive licenses, except for
`slog-json`, which uses the MPL. And we plan to replace `slog`.