BREAKING: Improve column case handling (fixes #84)
This patch fixes a number of issues handling mixed-case column names:

- PostgreSQL upsets should now work.
- Exporting from BigQuery should now work.
- BigQuery should preserve case but ignore it when comparing.

In order to do this, however, we had to remove the magic BigQuery
support for "cleaning up" incompatible column names. This was simply too
hard to keep consistent everywhere. So if you want to load data into
BigQuery, you must first make sure the column names are valid and
unique according to BigQuery, perhaps using a tool like `scrubcsv
--clean-column-names`.