bigquery: Honor `NOT NULL` on import (fixes #45)
For imports that required a custom SELECT to fix up the data, we used to
lose `NOT NULL` declarations on the columns, because we used `SELECT` to
create the destination table, and there's no way to represent `NOT
NULL`.

This patch rips all that out and replaces it with multi-statement SQL
scripts that include explicit `CREATE TABLE` statements.

This cleans up a bunch of dodgy logic in `bigquery::write_remote_data`
that I didn't really like. The whole business with `if_exists` and
upserts is now less hacksh.