gs,bigquery: Port to native Rust
We replace the `gsutil` and `bq` tools with native Rust code that
makes REST API calls directly. Consequences of this include:

1. We no longer require Google Cloud SDK for the regular application
   (but we still need it for tests)
2. Authentication only works by setting GCLOUD_SERVICE_ACCOUNT_KEY
   to service account key JSON.
3. Performance of Google Cloud Storage uploads and downloads may
   have decreased, and reliability may not be as high as before.

There are also a few TODO messages in dbcrossbarlib/src/clouds/gcloud
that need further investigation. But this is enough to get the
integration tests passing.

For use with a service account, I believe that you'll need the
following high-level permission:

- BigQuery Data Editor
- BigQuery Job User
- BigQuery User
- Storage Object Admin

The actual permissions you'll need are probably finer-grained than
this.