gs: Use parallel, chunked downloads (fixes #103)
Previously, we downloaded `gs://` objects using a single HTTP
stream. This had a number of drawbacks, including:

1. Backpressure from output drivers like BigML could cause commits to
   stall mid-download, sometimes for 15 minutes or more. To fix this
   we must commit to finishing any download streams we start. But this
   means we must download chunks small enough to fit in memory.
2. Various Google docs and rumors suggest that we can get better
   download performance by downloading ranges of a file in parallel,
   and assembling them locally.

So we reimplement our downloader to use `Range:` headers and the
`buffered` function to download multiple chunks in parallel but output
them in the correct order.

I haven't benchmarked this, but I hope it will at least match the
performance of `gsutil cp $URL -` in streaming mode, and use many
gigabytes less RAM.