schema+bigquery: Add support for STRUCT types
This is a change to `dbcrossbar`'s fundamental data model, adding
support for first-class structs with named fields. The ultimate goal
here is to support loading data directly into BigQuery STRUCT fields
instead of loading into STRING fields that contain serialized JSON.

This involves a lot complicated and hairy changes to the BigQuery
driver. This complexity is driven, in part, by the fact that BigQuery
has multiple data representations (scalars, arrays, structs,
stringified JSON, etc.), and the built-in import and export functions
do not necessarily handle this is a particularly consistent way. So we
basically need to build two mini "compilers" that generate custom JSON
UDFs for importing and exporting, and enforce our own idea of
consistency. This is made all the more complex because `dbcrossbar`
and BigQuery both support a wide variety of data types that don't
quite line up cleanly.

One of the big tricks we use to make this work is to create a new
`BqNonArrayDataType::Stringified` type, which serves two masters. As
far as BigQuery itself is concerned, it's just a `STRING`. As far as
`dbcrossbar` is concerned, it's still a portable `DataType`, but
represented as a string. This allows us to handle the weirder import
and export cases.

Many thanks to Prithaj for helping write the main new unit test that
verifies basic STRUCT import and export for a wide variety of types.

Co-authored-by: Prithaj Nath