Update to latest `peg` parser
We're going to add a new parser to `dbcrossbar` shortly, in order to
parse TypeScript-like schemas for our Shopify support. But before we
write the grammar for the new parser, we're going to update to the
latest `peg`.

`peg` implements "parser expression grammars", which are essentially a
bit like regular expressions with named, hierarchical rules. They're
much easier to create than classic LALR(1) grammars.

`peg` used to be implemented as a pre-processor that generated Rust
code, but it has now been ported to use Rust macros. This allows
grammars to be defined inside ordinary Rust files.

The new `peg` introduces many new features, but it removes two that we
relied upon: Negative character classes, and case-insensitive
literals.

We can implement negative character classes by replacing `[^\n]` (for
example) with `(![\n][_])`. The `![\n]` is a zero-width lookahead that
only matches if the next character is not `\n`, and `[_]` matches any
character. Together, they give us the old behavior in a slightly
uglier form.

To implement case-insensitive literals, we copy `rule i(literal:
&'static str)` from the `peg` issue tracker. This allows us to replace
the old `"CREATE"i` with `i("CREATE")`.