Implement a basic Shopify input driver
This uses the Shopify REST API and an accessToken to download JSON
data and convert it to CSV with embedded structs. This can be used as
follows:

    dbcrossbar cp \
      --if-exists=upsert-on:id \
      --temporary=gs://$TEMP_BUCKET \
      --temporary=bigquery:$PROJECT:$TEMP_DATASET \
      --schema='dbcrossbar-ts:shopify.ts#Order' \
      shopify://$SHOP/admin/api/2020-04/orders.json?status=any \
      bigquery:$PROJECT:$DATASET:$DEST_TABLE

This patch also includes a fix for `null` values in JSON where
optional arrays or structs are expected, which occurs frequently with
the Shopify API. And we update `shopify.ts` to make most fields
nullable, with the exception of `id` fields (which are useful for
upserts).