bigquery: Optionally specifiy job labels (fixes #123)
Google Cloud allows specifying optional labels on many kinds of
resources, in order to help keep track of costs.

This implements basic labels for BigQuery jobs, but not for the tables
we create.

We use our standard `--to-args` and `--from-args` syntax for the
specify the key/value pairs for each job label, instead of the Google
preferred `key:value` syntax. This will keep the `dbcrossbar` CLI more
consistent, and it makes it more obvious that `key` must be unique. It
also allows us to reuse the driver argument parsing code.