gs: Verify generations & CRC32C checksums
We add support for the `ifGenerationMatch` parameter across our Google
Cloud Storage code. This allows us to (1) make sure nobody has created
a file behind our back mid-operation, and (2) make sure that nobody
has replaced a file we're using.

We also check CRC32C checksums on upload. Ideally, we'd do this on
download, too, but we haven't implemented that.

This code was part of an attempt to add support for streaming
composite uploads to Google Cloud Storage, but this involves several
annoying sub-problems that I don't want to get into before the 0.4.0
release.

There's some more composite-related code that isn't included in this
PR. (I may check it into a WIP branch.) This explains why a few APIs
appear to have been generalized slightly for no obvious reason.