postgres: Stop replacing `jsonb[]` with `jsonb`
In general, `dbcrossbar` tries to faithfully transform the input schema
into an output schema, within the limitations of the databases and the
portable schema representation.

There has been one deliberate exception to this: We've replaced
`jsonb[]` with `jsonb` originally on BigQuery output, and now on
Postgres input. We've decided that this "opinionated" schema editing
has to go,  because every other time we've tried something like this,
we've been bitten by it.

Fixes #129.