gcloud: Try to use correct GCloud creds
We were caching OAuth2 tokens all in the same file. This meant that if
a user switched between different sets of credentials, we'd use the
cached creds from an earlier account.

We fix this by storing cached creds in different files. We try to find
cred-specific keys that we hash to build the file names.