postgres: Remove `diesel` (fixes #148)
We have historically used two Postgres drivers:

- `tokio-postgres` because it's async and supports bulk COPY commands
  nicely.
- `diesel` for more complex queries, because it had type checking.

The `diesel` code mostly predates `dbcrossbar`. We inherited it from
`schemaconv`. But it has a number of issues:

1. It requires static libpq and OpenSSL.
2. It's synchronous.
3. It has subtle different URL connection parameters than
   `tokio-postgres`.
4. It will never support `rustls`, allowing us to drop OpenSSL
   everywhere.

Furthermore, diesel didn't protect us from a type-related behavior
change in PostgreSQL 12. So I've decided to remove `diesel` and write a
small amount of raw SQL.

Of course, this means we lose compile-time checking. But since we don't
control `information_schema.*` or `Find_SRID`, that type checking doen't
buy us much.

Technically this breaks the `dbcrossbarlib` API, but since we don't
commit to semvar for that API yet, I'm not going to bump the version
accordingly.