Add support for PostgreSQL `ENUM`s (and portable named types)
This is a fairly major change, and it breaks the output format for
`dbcrossbar-schema` (so this is a semver bump). Highlists of this
change include:

- Parsing and printing `CREATE TYPE name AS ENUM(...)` statements.
- Creating `ENUM`s in the database if they don't exist.  Wrapping the
- existing `dbcrossbar-schema` output format in `{ "named_types": [],
- "tables": [old_table_format] }`.

Internally, this requires:

- Passing portable `Schema` around instead of `Table` in most places,
  so that we have access to `named_types`.  - Adding `DataType::Named`
  to refer to a named type.  - Adding `DataType::OneOf` to refer to a
  string enumeration.  - Lots of changes to the `postgres-sql` and
  `postgres` drivers.

We designed `DataType::Named` and `DataType::OneOf` the way we did in
order to later support the following planned features:

- `dbcrossbar-ts` should be able to convert `type` and `interface`
  declarations into named types.  - We want to support PostgreSQL
  native structure types, as an alternative to encoding structures as
  `jsonb` options.

Note that `one_of` and `named` support is fairly limited. Some drivers
will always recursively resolve all named types (like BigQuery), but
PostgreSQL only accepts `one_of` and `named` if they correspond to a
`CREATE TYPE name AS ENUM(...)` statement. These restrictions will be
relaxed over time as we figure out how to handle these cases
correctly, but we want to start by allowing only those cases we know
how to handle cleanly.

Other interesting PostgreSQL changes include:

- `TableName` is now `PgName`, because it's used by types and tables.

Many thanks to Tom Caruso for pairing on a tricky set of tests and
test fixes, and to Forrest Wallace and Seamus Abshere for providing
feedback on the design.