redshift: Try to use VARCHAR(MAX)
In RedShift, TEXT is an alias for VARCHAR(255). We
want VARCHAR(MAX) instead.

But since RedShift only works because it
piggybacks off of our better-supported PostgreSQL
driver, we need to thread the VARCHAR(MAX) type
through the postgres_shared module, and error out
if it shows up in PostgreSQL-specific code.

This has not yet been tested against RedShift,
because I don't have access to an instance right
now. But once a test system is available, this
should get us close.